requirements.txt
notion-client
python-telegram-bot
requests
pyyaml
bibtexparser
